requirements.txt
pytest-check
webdriver-manager
selenium
requests
Appium-Python-Client
pytest-xdist
pytest-ordering